Fixes and additions
Started work for login / registration page for admins and customers / html page for login form.

Also did some fixes for file names that didn't match between the controller and template pages. project runs but I didn't get around to testing if it was saving equipment or contracts to the DB. Users was working.